[+BUGFIX] Session is not written when storeSessionData is called
Although we have a special check for storeSessionData being called
when we determine if the cookie should be set or not, we are doing the
check (way) to late in the bootstrapping. So now we manually call
setSessinCookie (but only once) whenever the storeSessionData is
called and a change to sesData is detected.

Change-Id: I5bc8a3a17559103c289b8c7979ec7424236e2a15